Update routines
Ajout de Pick(), Place() et Square().
changement de int a float pour la variable Displacementmm dans StepperZ

+++ Code_OpenCR/stepperZ.hpp
@@ -82,7 +82,7 @@ void homeZ()
 void stepperGoToPos(int goalmm ) // goalmm, 
 {
   // Todo : add software limits
-  int displacementmm = goalmm - currPosmm; 
+  float displacementmm = goalmm - currPosmm; 
   currPosmm += displacementmm;
   int nMicroSteps = abs((displacementmm / MM_PER_TURN) * STEPS_PER_TURN * MICRO_STEPS); // Take absolue value
   int direction;
